test/requirements.txt
argparse==1.2.1
attrdict==2.0.0
backports.ssl-match-hostname==3.5.0.1
py==1.4.31
pytest==2.9.2
requests==2.10.0
six==1.10.0
websocket-client==0.37.0
wsgiref==0.1.2
